requirements.txt
jira==3.10.5
tabulate==0.9.0
streamlit==1.55.0
pandas==2.3.3
seaborn>=0.13.0
matplotlib>=3.8.0
python-dotenv==1.1.0
